windows warning

+++ dbcconvert/dbcconvert.cpp
@@ -34,7 +34,7 @@ private:
 };
 
 template <typename T>
-static T from_sv(std::string_view str, size_t base = 10)
+static T from_sv(std::string_view str, int base = 10)
 {
     T result;
     std::from_chars(str.data(), str.data() + str.size(), result, base);
